feat: add rating composable that draws an arc and colors it based on its value

+++ app/src/main/java/com/andreolas/movierama/ui/components/details/reviews/ReviewItemCard.kt
@@ -74,7 +74,7 @@ fun ReviewItemCard(
                 Rating(
                     modifier = Modifier
                         .padding(top = 12.dp, start = 12.dp, end = 12.dp),
-                    rating = review.rating.toString(),
+                    rating = review.rating.toInt().toString(),
                 )
             }
         }
